move Bases, CigarUtils into util pkg
