+++ src/main/scala/org/hammerlab/guacamole/commands/VariantSupport.scala
@@ -29,7 +29,8 @@ import org.hammerlab.guacamole.loci.set.LociSet
 import org.hammerlab.guacamole.pileup.Pileup
 import org.hammerlab.guacamole.reads.{InputFilters, MappedRead, ReadLoadingConfigArgs}
 import org.hammerlab.guacamole.reference.ReferenceBroadcast
-import org.hammerlab.guacamole.{Bases, ReadSet, SparkCommand}
+import org.hammerlab.guacamole.util.Bases
+import org.hammerlab.guacamole.{ReadSet, SparkCommand}
 import org.kohsuke.args4j.{Argument, Option => Args4jOption}
 
 object VariantSupport {
