+++ src/main/scala/org/hammerlab/guacamole/commands/jointcaller/pileup_summarization/ReadSubsequence.scala
@@ -1,10 +1,10 @@
 package org.hammerlab.guacamole.commands.jointcaller.pileup_summarization
 
 import org.apache.spark.Logging
-import org.hammerlab.guacamole.Bases
 import org.hammerlab.guacamole.pileup.PileupElement
 import org.hammerlab.guacamole.reads.MappedRead
 import org.hammerlab.guacamole.reference.ContigSequence
+import org.hammerlab.guacamole.util.Bases
 
 /**
  * A sub-sequence of the bases sequenced by a MappedRead.
