+++ src/main/scala/org/hammerlab/guacamole/pileup/Pileup.scala
@@ -18,10 +18,10 @@
 
 package org.hammerlab.guacamole.pileup
 
-import org.hammerlab.guacamole.Bases
 import org.hammerlab.guacamole.reads.MappedRead
-import org.hammerlab.guacamole.reference.{ ContigSequence, ReferenceBroadcast }
-import org.hammerlab.guacamole.variants.{ Allele, Genotype }
+import org.hammerlab.guacamole.reference.{ContigSequence, ReferenceBroadcast}
+import org.hammerlab.guacamole.util.Bases
+import org.hammerlab.guacamole.variants.{Allele, Genotype}
 
 /**
  * A [[Pileup]] at a locus contains a sequence of [[PileupElement]] instances, one for every read that overlaps that
