+++ src/main/scala/org/hammerlab/guacamole/pileup/PileupElement.scala
@@ -18,10 +18,10 @@
 
 package org.hammerlab.guacamole.pileup
 
-import htsjdk.samtools.{ CigarElement, CigarOperator }
-import org.hammerlab.guacamole.reference.{ ContigSequence, ReferenceBroadcast }
-import org.hammerlab.guacamole.{ Bases, CigarUtils }
+import htsjdk.samtools.{CigarElement, CigarOperator}
 import org.hammerlab.guacamole.reads.MappedRead
+import org.hammerlab.guacamole.reference.ContigSequence
+import org.hammerlab.guacamole.util.CigarUtils
 import org.hammerlab.guacamole.variants.Allele
 
 import scala.annotation.tailrec
