+++ src/main/scala/org/hammerlab/guacamole/reference/ReferenceBroadcast.scala
@@ -6,8 +6,8 @@ import java.util.NoSuchElementException
 import htsjdk.samtools.reference.FastaSequenceFile
 import org.apache.spark.SparkContext
 import org.apache.spark.broadcast.Broadcast
-import org.hammerlab.guacamole.Bases
 import org.hammerlab.guacamole.loci.set.LociSet
+import org.hammerlab.guacamole.util.Bases
 
 import scala.collection.mutable
