+++ src/main/scala/org/hammerlab/guacamole/variants/Genotype.scala
@@ -18,12 +18,9 @@
 
 package org.hammerlab.guacamole.variants
 
-import com.esotericsoftware.kryo.io.{ Input, Output }
-import com.esotericsoftware.kryo.{ Kryo, Serializer }
-import org.bdgenomics.adam.util.PhredUtils
+import com.esotericsoftware.kryo.io.{Input, Output}
+import com.esotericsoftware.kryo.{Kryo, Serializer}
 import org.bdgenomics.formats.avro.GenotypeAllele
-import org.hammerlab.guacamole.Bases.BasesOrdering
-import org.hammerlab.guacamole.pileup.PileupElement
 
 /**
  * A Genotype is a sequence of alleles of length equal to the ploidy of the organism.
