+++ src/main/scala/org/hammerlab/guacamole/variants/ReferenceVariant.scala
@@ -18,8 +18,9 @@
 
 package org.hammerlab.guacamole.variants
 
-import org.bdgenomics.formats.avro.{ DatabaseVariantAnnotation, Contig, Variant }
-import org.hammerlab.guacamole.{ Bases, HasReferenceRegion }
+import org.bdgenomics.formats.avro.{Contig, DatabaseVariantAnnotation, Variant}
+import org.hammerlab.guacamole.HasReferenceRegion
+import org.hammerlab.guacamole.util.Bases
 
 /**
  * Base properties of a genomic change in a sequence sample from a reference genome
