+++ src/test/scala/org/hammerlab/guacamole/assembly/DeBrujinGraphSuite.scala
@@ -1,8 +1,7 @@
 package org.hammerlab.guacamole.assembly
 
-import org.hammerlab.guacamole.Bases
 import org.hammerlab.guacamole.util.TestUtil.Implicits._
-import org.hammerlab.guacamole.util.{AssertBases, GuacFunSuite, TestUtil}
+import org.hammerlab.guacamole.util.{AssertBases, Bases, GuacFunSuite, TestUtil}
 
 class DeBruijnGraphSuite extends GuacFunSuite {
