+++ src/test/scala/org/hammerlab/guacamole/commands/SomaticStandardCallerSuite.scala
@@ -19,12 +19,10 @@
 package org.hammerlab.guacamole.commands
 
 import org.hammerlab.guacamole.reference.ReferenceBroadcast
-import org.hammerlab.guacamole.util.TestUtil
-import org.hammerlab.guacamole.Bases
+import org.hammerlab.guacamole.util.{Bases, GuacFunSuite, TestUtil}
 import org.hammerlab.guacamole.filters.SomaticGenotypeFilter
 import org.hammerlab.guacamole.pileup.Pileup
 import org.hammerlab.guacamole.reads.MappedRead
-import org.hammerlab.guacamole.util.{GuacFunSuite, TestUtil}
 import org.scalatest.Matchers
 import org.scalatest.prop.TableDrivenPropertyChecks
