+++ src/test/scala/org/hammerlab/guacamole/commands/jointcaller/PileupStatsSuite.scala
@@ -1,10 +1,9 @@
 package org.hammerlab.guacamole.commands.jointcaller
 
-import org.hammerlab.guacamole.Bases
 import org.hammerlab.guacamole.commands.jointcaller.pileup_summarization.PileupStats
 import org.hammerlab.guacamole.pileup.Pileup
 import org.hammerlab.guacamole.reference.ReferenceBroadcast
-import org.hammerlab.guacamole.util.{GuacFunSuite, TestUtil}
+import org.hammerlab.guacamole.util.{Bases, GuacFunSuite, TestUtil}
 
 class PileupStatsSuite extends GuacFunSuite {
   val cancerWGS1Bams = Seq("normal.bam", "primary.bam", "recurrence.bam").map(
