+++ src/test/scala/org/hammerlab/guacamole/likelihood/LikelihoodUtil.scala
@@ -20,8 +20,7 @@ package org.hammerlab.guacamole.likelihood
 
 import org.apache.spark.SparkContext
 import org.bdgenomics.adam.util.PhredUtils
-import org.hammerlab.guacamole.Bases
-import org.hammerlab.guacamole.util.TestUtil
+import org.hammerlab.guacamole.util.{Bases, TestUtil}
 import org.hammerlab.guacamole.variants.{Allele, Genotype}
 
 /**
