+++ src/test/scala/org/hammerlab/guacamole/reference/ReferenceBroadcastSuite.scala
@@ -1,7 +1,6 @@
 package org.hammerlab.guacamole.reference
 
-import org.hammerlab.guacamole.Bases
-import org.hammerlab.guacamole.util.{AssertBases, GuacFunSuite, TestUtil}
+import org.hammerlab.guacamole.util.{AssertBases, Bases, GuacFunSuite, TestUtil}
 import org.scalatest.Matchers
 
 class ReferenceBroadcastSuite extends GuacFunSuite with Matchers {
