+++ src/test/scala/org/hammerlab/guacamole/reference/ReferenceUtilsSuite.scala
@@ -1,8 +1,6 @@
 package org.hammerlab.guacamole.reference
 
-import org.hammerlab.guacamole.Bases
-import org.hammerlab.guacamole.util.TestUtil
-import org.hammerlab.guacamole.util.{GuacFunSuite, TestUtil}
+import org.hammerlab.guacamole.util.{Bases, GuacFunSuite, TestUtil}
 
 class ReferenceUtilsSuite extends GuacFunSuite {
