+++ src/test/scala/org/hammerlab/guacamole/util/BasesSuite.scala
@@ -16,9 +16,9 @@
  * limitations under the License.
  */
 
-package org.hammerlab.guacamole
+package org.hammerlab.guacamole.util
 
-import org.scalatest.{ FunSuite, Matchers }
+import org.scalatest.{FunSuite, Matchers}
 
 class BasesSuite extends FunSuite with Matchers {
