+++ src/test/scala/org/hammerlab/guacamole/util/TestUtil.scala
@@ -32,7 +32,7 @@ import org.hammerlab.guacamole.pileup.Pileup
 import org.hammerlab.guacamole.reads.{InputFilters, MappedRead, MateAlignmentProperties, PairedRead, Read, ReadLoadingConfig}
 import org.hammerlab.guacamole.reference.ReferenceBroadcast.MapBackedReferenceSequence
 import org.hammerlab.guacamole.reference.{ContigSequence, ReferenceBroadcast}
-import org.hammerlab.guacamole.{Bases, ReadSet}
+import org.hammerlab.guacamole.ReadSet
 import org.scalatest._
 
 import scala.collection.mutable
